requirements.txt
PySide2
pyensembl
biopython==1.76
numpy==1.18.1
